LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/view-timeline-keyframe-boundary-interpolation-expected.txt


FAIL ViewTimeline with timeline offset keyframes outside [0,1] assert_approx_equals: translation at contain 0% expected 100 +/- 0.000001 but got 0
